+++ lib/include/aul_internal.h
@@ -9,6 +9,7 @@
  * @copyright Copyright (c) 2026
  */
 
+#define AUL_INTERNAL
 #ifndef _AUL_INTERNAL_H_
 #define _AUL_INTERNAL_H_
 #ifdef AUL_INTERNAL
@@ -33,6 +34,22 @@ struct aurora_user_library_context {
 };
 
 extern struct aurora_user_library_context _aul_ctx;
+
+#define TIME_REGION(name)                                                      \
+    for (struct {                                                              \
+             struct timespec start;                                            \
+             int done;                                                         \
+         } _t = {{0}, 0};                                                      \
+         !_t.done && (clock_gettime(CLOCK_MONOTONIC, &_t.start), 1);           \
+         __extension__({                                                       \
+             struct timespec end;                                              \
+             clock_gettime(CLOCK_MONOTONIC, &end);                             \
+             double ms =                                                       \
+                 (double) (end.tv_sec - _t.start.tv_sec) * 1000.0 +            \
+                 (double) (end.tv_nsec - _t.start.tv_nsec) / 1000000.0;        \
+             log_info("[TIMER] %-20s : %.6f ms", name, ms);                    \
+             _t.done = 1;                                                      \
+         }))
 
 #endif
 #endif
