+++ tests/discovery/test_discovery.c
@@ -43,6 +43,7 @@ int main(int argc, char **argv) {
         buf[i] = 0;
     }
     AUL_Mem_unprotect(3);
+    s = AUL_Checkpoint(1, "TestCkpt0000001");
 
     printf("Starting Restore\n");
     // *nowarn*
